requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20200713
distlib==0.3.1
Django==1.11.6
django-cors-headers==3.5.0
django-crispy-forms==1.6.0
django-filter==1.0.4
django-formtools==1.0
django-import-export==0.5.1
django-redis==4.12.1
django-reversion==2.0.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
docx==0.2.4
drf-extensions==0.6.0
filelock==3.0.12
future==0.15.2
httplib2==0.9.2
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
itypes==1.2.0
Jinja2==2.11.2
lxml==4.5.1
Markdown==2.6.8
MarkupSafe==1.1.1
mysqlclient==1.4.0
pbr==5.4.5
pefile==2019.4.18
Pillow==7.1.2
PyAudio==0.2.11
PyInstaller==3.6
PyJWT==1.7.1
python-docx==0.8.10
pytz==2020.1
pywin32-ctypes==0.2.0
redis==3.5.3
requests==2.24.0
six==1.10.0
sqlparse==0.3.1
stevedore==3.2.0
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
XlsxWriter==1.3.3
xlwt==1.3.0
zipp==3.1.0
